requirements.txt
fastapi
uvicorn[standard]
torch
transformers
pydantic
httpx
asyncio
python-dotenv
aiohttp